+++ src/z80asm/hist.c
@@ -9,8 +9,8 @@ Repository: https://github.com/pauloscustodio/z88dk-z80asm
 
 #include "hist.h"
 
-#define VERSION     "2.8.2"
-#define COPYRIGHT   "InterLogic 1993-2009, Paulo Custodio 2011-2015"
+#define VERSION     "2.8.4"
+#define COPYRIGHT   "InterLogic 1993-2009, Paulo Custodio 2011-2016"
 
 #ifdef QDOS
 #include <qdos.h>